Fix leaderboard loading and improved metrics

+++ crates/frontend/src/services/leaderboard.rs
@@ -1,7 +1,6 @@
 use super::requests::request_get;
 use crate::error::Error;
 
-/// Get decks filtered by author
-pub async fn get_leaderboard() -> Result<Vec<(String, i32)>, Error> {
-    request_get::<Vec<(String, i32)>>("/leaderboard".to_string()).await
+pub async fn get_leaderboard() -> Result<Vec<(String, usize)>, Error> {
+    request_get::<Vec<(String, usize)>>("/leaderboard".to_string()).await
 }
